requirements.txt
flask
waitress
infi.systray